+++ home/views.py
@@ -7,10 +7,14 @@ from programs import http_funcs, log_funcs, ap, version
 
 # Create your views here.
 def show(request):
+    user_id = request.COOKIES.get("anonymous_user_id")
+    print(f"{user_id = }")
+    if user_id:
+        http_funcs.set_user_sql(request, models.User, user_id)
     if http_funcs.is_ajax(request):
-        # 写数据表
-        fingerprint = json.loads(request.body.decode('utf-8'))['fingerprint']
-        http_funcs.set_user_sql(request, models.User, fingerprint)
+        # # 写数据表
+        # fingerprint = json.loads(request.body.decode('utf-8'))['fingerprint']
+        # http_funcs.set_user_sql(request, models.User, fingerprint)
         return JsonResponse({})
     else:
         log_funcs.write_log(http_funcs.get_ip(request), 'info', 'Visit home html')