+++ programs/__init__.py
@@ -1,9 +1,9 @@
 """
 Import ararpy
 """
-version = "20251001"
-# import ararpy as ap                 # using Pip version
-try:
-    from .local_init import *
-except ImportError:
-    import ararpy_package.ararpy as ap    # local
+version = "20251202"
+import ararpy as ap                 # using Pip version
+# try:
+#     from .local_init import *
+# except ImportError:
+#     import ararpy_package.ararpy as ap    # local